requirements.txt
attrdict==2.0.1
moviepy==1.0.3
pybullet==3.2.6
Pygments==2.19.1
easydict==1.13
matplotlib==3.8.3
numpy
opencv_python_headless==4.8.1.78
Pillow==11.0.0
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.4.1.post1
supervision
transformers==4.38.1
open3d
scipy==1.15.2
trimesh==4.6.4
seaborn==0.13.2